pyproject.toml
[project]
name = "oasis-dashboard"
version = "1.1.0"
description = "OASIS Dashboard - A monitoring and management dashboard for OASIS framework"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "camel-oasis>=0.2.0",
    "datasets>=4.8.4",
    "neo4j>=5.0.0",
    "neo4j-driver>=5.28.3",
    "pymongo>=4.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oasis_dashboard"]

[dependency-groups]
dev = []
